requirements.txt
gradio
torch
torchvision
Pillow
gdown
numpy
scipy
cmake
onnxruntime-gpu
opencv-python-headless
encoded-video
basicsr
facexlib
gfpgan
ffmpeg-python
onnxruntime-gpu
pytorchvideo
timm